fix comments

+++ window_scene/session/host/src/session.cpp
@@ -4165,7 +4165,7 @@ WindowMetaInfo Session::GetWindowMetaInfoForWindowInfo() const
     windowMetaInfo.pid = GetCallingPid();
     windowMetaInfo.windowType = GetWindowType();
     if (auto parentSession = GetParentSession()) {
-        windowMetaInfo.parentWindowId = static_cast<uint32_t>(parentSession->GetWindowId());
+        windowMetaInfo.parentWindowId = parentSession->GetWindowId();
     }
     if (auto surfaceNode = GetSurfaceNode()) {
         windowMetaInfo.surfaceNodeId = static_cast<uint64_t>(surfaceNode->GetId());
